requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
decouple==0.0.7
distlib==0.3.3
Django==3.2.7
env==0.1.0
filelock==3.3.0
future==0.18.2
idna==3.3
line-bot-sdk==1.20.0
mysql-client==0.0.1
Pillow==8.4.0
platformdirs==2.4.0
pycodestyle==2.8.0
PyMySQL==1.0.2
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
virtualenv==20.8.1
